refactor: apply clippy lints across Rust backend and fix test cast
- Replace closures with function references and method calls
- Use idiomatic patterns: struct init, .abs_diff(), .or_default(), range.contains()
- Accept &Path instead of &PathBuf in function signatures
- Replace `if let Ok(_)` with `.is_ok()` and use `is_multiple_of()`
- Simplify Option chains with .map() instead of if-let-else
- Replace `as any` with `as never` in test to satisfy lint rules

+++ apps/desktop/src-tauri/src/commands/capture_overlay/wndproc.rs
@@ -110,7 +110,7 @@ fn handle_nchittest(state_ptr: *mut OverlayState, lparam: LPARAM) -> LRESULT {
 fn handle_set_cursor(state_ptr: *mut OverlayState, lparam: LPARAM) -> LRESULT {
     unsafe {
         // Only handle cursor in client area
-        if (lparam.0 as u32 & 0xFFFF) != HTCLIENT as u32 {
+        if (lparam.0 & 0xFFFF) != HTCLIENT as isize {
             return LRESULT(0);
         }
